Authentification and authorization using jwt tokens

+++ CarPartsShop.API/Controllers/CarBrandController.cs
@@ -3,6 +3,11 @@ using Microsoft.AspNetCore.Mvc;
 using CarPartsShop.API.Models;
 using CarPartsShop.Core.Interfaces;
 using System.Collections.Generic;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
+using Microsoft.AspNetCore.Authorization;
+using System.Data;
+using CarPartsShop.Core.Aggregates.Auth;
 
 namespace CarPartsShop.API.Controllers;
 
@@ -10,14 +15,17 @@ namespace CarPartsShop.API.Controllers;
 [Route("api/carBrand")]
 public class CarBrandController : BaseController
 {
-    public CarBrandController(ICarBrandService carBrandService, ICarModelService carModelService)
+    private readonly IAuthorizationService _authorizationService;
+    public CarBrandController(ICarBrandService carBrandService, ICarModelService carModelService, ICarPartService carPartService)
     {
         CarBrandService = carBrandService ?? throw new ArgumentNullException(nameof(carBrandService));
         CarModelService = carModelService ?? throw new ArgumentNullException(nameof(carModelService));
+        CarPartService = carPartService ?? throw new ArgumentNullException(nameof(carPartService));
     }
 
     private ICarBrandService CarBrandService { get; }
     public ICarModelService CarModelService { get; }
+    public ICarPartService CarPartService { get; }
 
     [HttpGet]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarBrandModel))]
@@ -55,6 +63,7 @@ public class CarBrandController : BaseController
             return StatusCode(StatusCodes.Status500InternalServerError, ex);
         }
     }
+
     [HttpGet("{carBrandId:Guid}/carModel")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarBrandModel))]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -63,6 +72,42 @@ public class CarBrandController : BaseController
         try
         {
             var result = await CarModelService.GetCarModelByBrandList(carBrandId);
+            if (carBrandId == Guid.Empty) return NotFound();
+            if (result is null) return NotFound();
+            //return Mapper.Map<CarModelModel>(result);
+            return Ok(result);
+        }
+        catch (Exception ex)
+        {
+            return StatusCode(StatusCodes.Status500InternalServerError, ex);
+        }
+    }
+
+    [HttpGet("{carBrandId:Guid}/carModel/{carModelId:Guid}/carPart")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarBrandModel))]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<ActionResult<ICollection<CarPartModel>>> GetCarPartByBrandList([FromRoute] Guid carBrandId, [FromRoute] Guid carModelId, CancellationToken cancellationToken = default)
+    {
+        try
+        {
+            var result = await CarPartService.GetCarPartByBrandList(carBrandId, carModelId, cancellationToken);
+            if (carBrandId == Guid.Empty) return NotFound();
+            //return Mapper.Map<CarModelModel>(result);
+            return Ok(result);
+        }
+        catch (Exception ex)
+        {
+            return StatusCode(StatusCodes.Status500InternalServerError, ex);
+        }
+    }
+    [HttpGet("{carBrandId:Guid}/carPart")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarBrandModel))]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<ActionResult<ICollection<CarPartModel>>> GetCarPartsByBrandList([FromRoute] Guid carBrandId, CancellationToken cancellationToken = default)
+    {
+        try
+        {
+            var result = await CarPartService.GetCarPartsByBrandList(carBrandId, cancellationToken);
             if (carBrandId == Guid.Empty) return NotFound();
             //return Mapper.Map<CarModelModel>(result);
             return Ok(result);
@@ -73,16 +118,18 @@ public class CarBrandController : BaseController
         }
     }
     [HttpPost]
+    [Authorize(Roles = ShopRoles.ShopWorker)]
     [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(CarBrandModel))]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public async Task<ActionResult<CarBrandModel>> CreateCarBrand(CarBrandModel carBrandModel, CancellationToken cancellationToken = default)
+    public async Task<ActionResult<CarBrandModel>> CreateCarBrand([FromBody] CarBrandModel carBrandModel, CancellationToken cancellationToken = default)
     {
         try
         {
             if (carBrandModel is null) return BadRequest();
-
-            var createdCarBrand = await CarBrandService.CreateCarBrand(carBrandModel.ToEntity(), cancellationToken);
-
+            var carBrandEntity = carBrandModel.ToEntity();
+            carBrandEntity.UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
+            var createdCarBrand = await CarBrandService.CreateCarBrand(carBrandEntity, cancellationToken);
+            
             return Created(nameof(CarBrandModel), Mapper.Map<CarBrandModel>(carBrandModel));
             //return Ok(Mapper.Map<CarBrandModel>(carBrandModel));
         }
@@ -92,6 +139,7 @@ public class CarBrandController : BaseController
         }
     }
     [HttpDelete("{carBrandId:Guid}")]
+    [Authorize(Roles = ShopRoles.ShopWorker)]
     [ProducesResponseType(StatusCodes.Status204NoContent, Type = typeof(CarBrandModel))]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<CarBrandModel>> DeleteCarBrand(Guid carBrandId, CancellationToken cancellationToken = default)
@@ -101,6 +149,8 @@ public class CarBrandController : BaseController
             /*var carBrandToDelete = await CarBrandService.GetCarBrandById(carBrandId);
             var deletedcarBrand = await CarBrandService.DeleteCarBrand(carBrandToDelete, cancellationToken);*/
             if (carBrandId == Guid.Empty) return NotFound();
+            var itemToDelete = await GetCarBrandById(carBrandId);
+            if (itemToDelete is null) return NotFound();
             await CarBrandService.DeleteCarBrand(carBrandId, cancellationToken);
             return NoContent();
             //return Mapper.Map<CarBrandModel>(deletedcarBrand);
@@ -111,6 +161,7 @@ public class CarBrandController : BaseController
         }
     }
     [HttpPut("{carBrandId:Guid}")]
+    [Authorize(Roles = ShopRoles.ShopWorker)]
     [ProducesResponseType(StatusCodes.Status204NoContent, Type = typeof(CarBrandModel))]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
@@ -118,8 +169,15 @@ public class CarBrandController : BaseController
     {
         try
         {
+            var brand = CarBrandService.GetCarBrandById(carBrandId, cancellationToken);
             if (carBrandId != carBrand.Id) return BadRequest("Nesutampa ID");
             if (carBrandId == Guid.Empty) return NotFound();
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, brand, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
             await CarBrandService.UpdateCarBrand(carBrandId, carBrand.ToEntity(), cancellationToken);
             return NoContent();
         }
